restructured tests and fixed references
