+++ tests/unit/all.ts
@@ -1,7 +1,7 @@
-import './createButton';
 import './createProjector'; // projector should load before integrations, because it loads a shim
+import './createWidgetBase';
+import './d';
 import './integrations';
 import './main';
+import './components/all';
 import './mixins/all';
-import './bases/all';
-import './util/all';
