+++ tests/unit/components/all.ts
@@ -1,0 +1,2 @@
+import './button/createButton';
+import './textinput/createTextInput';
