+++ tests/unit/createWidgetBase.ts
@@ -1,8 +1,8 @@
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
-import createWidgetBase from '../../../src/bases/createWidgetBase';
+import createWidgetBase from '../../src/createWidgetBase';
 import { DNode, HNode, WidgetState, WidgetOptions } from 'dojo-interfaces/widgetBases';
-import d from '../../../src/util/d';
+import d from '../../src/d';
 import { stub } from 'sinon';
 
 registerSuite({
