--- tests/unit/util/all.ts
@@ -1,1 +1,0 @@
-import './d';
